Split vultron/core/behaviors/sync/nodes.py into subpackage by semantic domain
Issue #754: Refactor flat nodes.py (693 lines, 15 node classes) into a nodes/
subpackage grouped by semantic concern, following BTND-07-001/002.

Changes:
- vultron/core/behaviors/sync/nodes.py → nodes/ subpackage:
  * conditions.py: 5 condition nodes (Check* and Verify*) + helpers
  * receive.py: 3 action nodes (receive/validate log entries)
  * chain.py: 4 action nodes (chain reconstruction, entry creation)
  * replay.py: 3 action nodes (replay, fan-out, find case actor)
  * __init__.py: re-exports all public names for import compatibility

- test/core/behaviors/sync/:
  * test_nodes.py → nodes/conftest.py: shared fixtures
  * Added test_conditions.py, test_receive.py, test_chain.py, test_replay.py
  * Public import paths preserved (from sync.nodes import ...)

All 3006 tests pass. Node imports in embargo/nodes.py and tree files
continue to work via __init__.py re-exports.
