+++ test/core/behaviors/sync/nodes/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/env python
+"""Unit tests for sync BT node submodules."""
