+++ test/core/behaviors/sync/nodes/test_replay.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+"""Unit tests for sync replay nodes."""
+
+
+def test_replay_nodes_placeholder():
+    """Placeholder test for replay nodes. Full coverage in integration tests."""
+    pass
